% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRFSS19952010_data.R
\docType{data}
\name{BRFSS19952010}
\alias{BRFSS19952010}
\title{Pooled 1995-2010 BRFSS data}
\format{This data set should cover years 1995-2010 and individuals aged over 65
\describe{
    \item{MEDCOST}{Have troubles to see doctors because of cost}
}}
\source{
\url{https://www.cdc.gov/brfss/index.html}
}
\usage{
BRFSS19952010
}
\description{
This is the BRFSS data after reading, merging, and cleaning.
We use this data set in regression analysis.
(and possibly for the purpose of summary statistics and descriptive analysis)
}
\examples{
data(BRFSS19952010)
}
\keyword{datasets}
