% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_BRFSS.R
\name{merge_BRFSS}
\alias{merge_BRFSS}
\title{Merge the list of the BRFSS data frames into a single data frame}
\usage{
merge_BRFSS(df_list)
}
\arguments{
\item{df_list}{A list of BRFSS data frames}
}
\value{
A data frame
}
\description{
Use this function to integrate the BRFSS data frames into a single data frame
}
\examples{
lapply(1995:2010, read_BRFSS) \%>\% merge_BRFSS()

}
\author{
Jiacheng He
}
