% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msp.R
\name{msp}
\alias{msp}
\title{Fix misspelling and different in similar character data.}
\usage{
msp(x, y, r = 0.8, type = "ask", ...)
}
\arguments{
\item{x}{a character to search}

\item{y}{the vector containing x and order character}

\item{r}{level of similarity between element, by default is 0.8}

\item{type}{an charactaer indicating the remplacement procudure. Se details for mor information}

\item{...}{additional arguments are passed on to \link{stringsim} function.}
}
\value{
a same object than \code{x} without misspelling
}
\description{
Look for similatiries between a caharacter and a vector character to detect misspeling problems and change it.
}
\details{
Type define the way of the rigth option to replace is choose,  |code{ask} generate a interactive replacement process, \code{first} use the fisrt element of the list as correct element to repalce and \code{long} use the longer element as correct to replace. The default value is \code{ask}
}
\examples{
\dontrun{
	set.seed(1)
	z <- stringi::stri_rand_strings(20, 5)
	a <- z[1]
	a1 <- c("GNZuC","GNZuC ", "GNZu.C"," GNZuC", "GNZ uC", "gNZuC")
	z <- c(z,sample(z),sample(z,10),a1)
	msp(a,z)
}
}
