% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BRFSS.R
\name{read_BRFSS}
\alias{read_BRFSS}
\title{Loading BRFSS data}
\usage{
read_BRFSS(year)
}
\arguments{
\item{year}{}
}
\value{
A data frame
}
\description{
Specify the year of the BRFSS and load it as a data frmae
}
\examples{
BRFSS1995 <- read_BRFSS(1995)

}
\author{
Jiacheng He
}
