% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{specification}
\alias{specification}
\title{Specify the regression model}
\usage{
specification(LHS, model, controls = FALSE)
}
\arguments{
\item{LHS}{The LHS outcome variable name}

\item{model}{Model type: "DID" or "event_study"}

\item{controls}{Whether to include controls}
}
\value{
An expression of the regression specification
}
\description{
Specify the regression model
}
\examples{

}
\author{
Jiacheng He
}
