% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_regression.R
\name{subgroup_regression}
\alias{subgroup_regression}
\title{Select the subgroup of the data and perform regression analysis}
\usage{
subgroup_regression(condition, df, model, cluster = NULL,
  event_plot = FALSE, xlab = NULL, ylab = NULL)
}
\arguments{
\item{condition}{The condition to subset the data. A character.}

\item{df}{The data frame that contains the regression variables}

\item{model}{The regression model}

\item{cluster}{Whether to cluster the standard error}

\item{event_plot}{Whether to generate the event study plot}

\item{xlab}{The xlabel of the event plot}

\item{ylab}{The ylabel of the event plot}
}
\value{
A list of regression result (and an event plot if specified)
}
\description{
Heterogeneity analysis
}
\examples{

}
\author{
Jiacheng He
}
